🗺️ Atlas: Move TxVisibilityManager to core to break db/api dependency tangle
