+++ src/api/transaction/mod.rs
@@ -61,13 +61,13 @@
 
 pub mod read_tx;
 pub mod types;
-pub mod visibility;
+
 pub mod write;
 pub mod write_buffer;
 
 pub use read_tx::ReadTransaction;
 pub use types::{TxId, TxMetadata, TxState};
-pub use visibility::{CompressionStats, TransactionSnapshot, TxVisibilityManager};
+
 pub use write::WriteTransaction;
 pub use write_buffer::{BufferedWrite, WriteBuffer};
