+++ src/db/mod.rs
@@ -3,9 +3,9 @@
 //! This module provides the primary interface to the database, coordinating
 //! between current storage (fast path) and historical storage (temporal path).
 
-use crate::api::transaction::TxVisibilityManager;
 use crate::core::id::{IdGenerator, TxIdGenerator};
 use crate::core::temporal::Timestamp;
+use crate::core::transaction_visibility::TxVisibilityManager;
 use crate::index::temporal::TemporalIndexes;
 use crate::query::planner::Statistics;
 use crate::storage::current::CurrentStorage;
